requirements.txt
chromadb==0.4.24
elasticsearch7==7.17.9
jieba==0.42.1
nltk==3.8.1
numpy==1.26.4
openai==1.13.3
pdfminer.six==20231228
python-dotenv==1.0.1
Requests==2.31.0
sentence_transformers==2.2.2
